+++ CHANGELOG.md
@@ -12,6 +12,7 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 ## [Unreleased]
 
 - [OS-58] New company address fields
+- Custom permissions by term field
 
 ## [3.10.0] 2023-08-23